refactor: release filter
